First steps, awww

+++ scripts/cleanup.sh
@@ -1,4 +1,17 @@
-rm -f scratch* 
-rm -f slurm* 
-rm -f STD*
-rm -f tapes*
+#!/bin/bash
+# cleanup after a run
+
+read -p "Are you sure you want to clean up this directory? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    # clean up directory 
+    rm -f scratch*
+    cp STDOUT.0000 stdOUT.txt
+    cp STDERR.0000 stdERR.txt
+    rm -f STDOUT* STDERR*
+    rm -rf .0*
+    rm -rf tapes/
+    rm -f tapes*
+    rm slurm*
+fi
